+++ lib/test/apyfloatarray/test_methods.py
@@ -996,3 +996,33 @@ def test_ravel(shape):
     reshaped = arr.reshape(shape)
     if not APyFloatArray.is_identical(reshaped.ravel(), arr):
         pytest.fail(f"Flatten didn't return to original 1d list after reshape {shape}")
+
+
+def test_swapaxes():
+    signs = [1, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 0]
+    exps = [5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60]
+    mans = [3, 1, 4, 2, 6, 5, 8, 7, 9, 0, 2, 3]
+
+    a = APyFloatArray(
+        signs=signs, exps=exps, mans=mans, exp_bits=5, man_bits=2
+    ).reshape((6, 2))
+
+    if not a.swapaxes(0, 1).is_identical(a.T):
+        pytest.fail("swapaxes didn't correctly swap axis")
+
+    if not a.swapaxes(1, 0).is_identical(a.T):
+        pytest.fail("swapaxes didn't correctly swap axis")
+
+    a = APyFloatArray.from_float([0] * 24, exp_bits=5, man_bits=2).reshape((4, 3, 2))
+
+    if not a.swapaxes(0, 1).shape == (3, 4, 2):
+        pytest.fail("swapaxes didn't correctly swap axis")
+
+    if not a.swapaxes(1, 0).shape == (3, 4, 2):
+        pytest.fail("swapaxes didn't correctly swap axis")
+
+    if not a.swapaxes(2, 0).shape == (2, 3, 4):
+        pytest.fail("swapaxes didn't correctly swap axis")
+
+    if not a.swapaxes(0, 2).shape == (2, 3, 4):
+        pytest.fail("swapaxes didn't correctly swap axis")