+++ src/apyfixedarray.cc
@@ -731,6 +731,29 @@ APyFixedArray APyFixedArray::matmul(const APyFixedArray& rhs) const
     ));
 }
 
+APyFixedArray APyFixedArray::swapaxes(nb::int_ axis1, nb::int_ axis2) const
+{
+    size_t _axis1 = ::get_normalized_axes(axis1, _ndim).front();
+    size_t _axis2 = ::get_normalized_axes(axis2, _ndim).front();
+
+    std::vector<size_t> new_axis(_ndim);
+    std::iota(new_axis.begin(), new_axis.end(), 0);
+
+    // Swap the specified axes
+    std::swap(new_axis[_axis1], new_axis[_axis2]);
+
+    std::vector<size_t> new_shape(_ndim);
+    for (size_t i = 0; i < _ndim; ++i) {
+        new_shape[i] = _shape[new_axis[i]];
+    }
+
+    APyFixedArray result(new_shape, bits(), int_bits());
+    ::transpose_axes_and_copy_data(
+        _data.begin(), result._data.begin(), _shape, new_axis, _itemsize
+    );
+    return result;
+}
+
 APyFixedArray APyFixedArray::transpose(std::optional<nb::tuple> axes) const
 {
     switch (_ndim) {